create CommonUtils for commonly used methods
